Track applied plugins only once
We were tracking plugins both in the plugin
manager as well as in the plugin container.
The container was behaving like any other of
our domain object containers, which is not quite
right, because users cannot add/remove plugins
manually. It now forbids these operations and
uses the store from the plugin manager directly
instead of keeping another copy of it in memory.