Calculate the C/C++ source files of a compile task once only during execution, rather than twice (when outputs can be reused from somewhere) or 3 times (when the task action is executed).
Added an initial rough implementation of a `ConfigurableFileCollection` that can be used as a task input property, and which caches the matching files during task execution. This means that these input files are queried once during execution regardless of whether the file collection is used to calculate other input values, such as the header files for a C/C++ compile task, or used by the task action when it runs (which it always will be). This is intended to be evolved into a reusable implementation but currently lives with the native compile tasks.

Also changed the native compile tasks to discard incremental compile analysis state at the end of task execution rather than retain it in memory for the remainder of the build.